requirements.txt
discord.py[voice]
psycopg2